requirements.txt
numpy
pathlib
pymel
GitPython